+++ pkg/operator/ceph/cluster/mon/mon.go
@@ -35,12 +35,9 @@ import (
 	cephconfig "github.com/rook/rook/pkg/daemon/ceph/config"
 	mondaemon "github.com/rook/rook/pkg/daemon/ceph/mon"
 	"github.com/rook/rook/pkg/operator/k8sutil"
-	"github.com/rook/rook/pkg/util"
-	v1 "k8s.io/api/core/v1"
+	"k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/intstr"
-	"k8s.io/kubernetes/pkg/kubelet/apis"
 )
 
 var logger = capnslog.NewPackageLogger("github.com/rook/rook", "op-mon")
@@ -121,8 +118,16 @@ type NodeInfo struct {
 }
 
 // New creates an instance of a mon cluster
-func New(context *clusterd.Context, namespace, dataDirHostPath, rookVersion string, cephVersion cephv1.CephVersionSpec, mon cephv1.MonSpec,
-	placement rookalpha.Placement, hostNetwork bool, resources v1.ResourceRequirements, ownerRef metav1.OwnerReference) *Cluster {
+func New(
+	context *clusterd.Context,
+	namespace, dataDirHostPath, rookVersion string,
+	cephVersion cephv1.CephVersionSpec,
+	mon cephv1.MonSpec,
+	placement rookalpha.Placement,
+	hostNetwork bool,
+	resources v1.ResourceRequirements,
+	ownerRef metav1.OwnerReference,
+) *Cluster {
 	return &Cluster{
 		context:              context,
 		placement:            placement,
@@ -230,7 +235,11 @@ func (c *Cluster) initMonConfig(size int) []*monConfig {
 
 	// initialize the mon pod info for mons that have been previously created
 	for _, monitor := range c.clusterInfo.Monitors {
-		mons = append(mons, &monConfig{ResourceName: resourceName(monitor.Name), DaemonName: monitor.Name, Port: getPortFromEndpoint(monitor.Endpoint)})
+		mons = append(mons, &monConfig{
+			ResourceName: resourceName(monitor.Name),
+			DaemonName:   monitor.Name,
+			Port:         getPortFromEndpoint(monitor.Endpoint),
+		})
 	}
 
 	// initialize mon info if we don't have enough mons (at first startup)
@@ -244,7 +253,11 @@ func (c *Cluster) initMonConfig(size int) []*monConfig {
 
 func newMonConfig(monID int) *monConfig {
 	daemonName := k8sutil.IndexToName(monID)
-	return &monConfig{ResourceName: resourceName(daemonName), DaemonName: daemonName, Port: int32(mondaemon.DefaultPort)}
+	return &monConfig{
+		ResourceName: resourceName(daemonName),
+		DaemonName:   daemonName,
+		Port:         int32(mondaemon.DefaultPort),
+	}
 }
 
 // resourceName ensures the mon name has the rook-ceph-mon prefix
@@ -275,50 +288,6 @@ func (c *Cluster) initMonIPs(mons []*monConfig) error {
 	}
 
 	return nil
-}
-
-func (c *Cluster) createService(mon *monConfig) (string, error) {
-	labels := c.getLabels(mon.DaemonName)
-	s := &v1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:   mon.ResourceName,
-			Labels: labels,
-		},
-		Spec: v1.ServiceSpec{
-			Ports: []v1.ServicePort{
-				{
-					Name:       mon.ResourceName,
-					Port:       mon.Port,
-					TargetPort: intstr.FromInt(int(mon.Port)),
-					Protocol:   v1.ProtocolTCP,
-				},
-			},
-			Selector: labels,
-		},
-	}
-	k8sutil.SetOwnerRef(c.context.Clientset, c.Namespace, &s.ObjectMeta, &c.ownerRef)
-	if c.HostNetwork {
-		s.Spec.ClusterIP = v1.ClusterIPNone
-	}
-
-	s, err := c.context.Clientset.CoreV1().Services(c.Namespace).Create(s)
-	if err != nil {
-		if !errors.IsAlreadyExists(err) {
-			return "", fmt.Errorf("failed to create mon service. %+v", err)
-		}
-		s, err = c.context.Clientset.CoreV1().Services(c.Namespace).Get(mon.ResourceName, metav1.GetOptions{})
-		if err != nil {
-			return "", fmt.Errorf("failed to get mon %s service ip. %+v", mon.ResourceName, err)
-		}
-	}
-
-	if s == nil {
-		logger.Warningf("service ip not found for mon %s. this better be a test", mon.ResourceName)
-		return "", nil
-	}
-
-	logger.Infof("mon %s running at %s:%d", mon.DaemonName, s.Spec.ClusterIP, mon.Port)
-	return s.Spec.ClusterIP, nil
 }
 
 func (c *Cluster) assignMons(mons []*monConfig) error {
@@ -353,25 +322,6 @@ func (c *Cluster) assignMons(mons []*monConfig) error {
 
 	logger.Debug("mons have been assigned to nodes")
 	return nil
-}
-
-func getNodeInfoFromNode(n v1.Node) (*NodeInfo, error) {
-	nr := &NodeInfo{
-		Name:     n.Name,
-		Hostname: n.Labels[apis.LabelHostname],
-	}
-
-	for _, ip := range n.Status.Addresses {
-		if ip.Type == v1.NodeExternalIP || ip.Type == v1.NodeInternalIP {
-			logger.Debugf("using IP %s for node %s", ip.Address, n.Name)
-			nr.Address = ip.Address
-			break
-		}
-	}
-	if nr.Address == "" {
-		return nil, fmt.Errorf("couldn't get IP of node %s", nr.Name)
-	}
-	return nr, nil
 }
 
 func (c *Cluster) startDeployments(mons []*monConfig, index int) error {
@@ -451,95 +401,6 @@ func (c *Cluster) saveMonConfig() error {
 	}
 
 	return nil
-}
-
-// getMonNodes detects the nodes that are available for new mons to start.
-func (c *Cluster) getMonNodes() ([]v1.Node, error) {
-	availableNodes, nodes, err := c.getAvailableMonNodes()
-	if err != nil {
-		return nil, err
-	}
-	logger.Infof("Found %d running nodes without mons", len(availableNodes))
-
-	// if all nodes already have mons and the user has given the mon.count, add all nodes to be available
-	if c.AllowMultiplePerNode && len(availableNodes) == 0 {
-		logger.Infof("All nodes are running mons. Adding all %d nodes to the availability.", len(nodes.Items))
-		for _, node := range nodes.Items {
-			valid, err := k8sutil.ValidNode(node, c.placement)
-			if err != nil {
-				logger.Warning("failed to validate node %s %v", node.Name, err)
-			} else if valid {
-				availableNodes = append(availableNodes, node)
-			}
-		}
-	}
-
-	return availableNodes, nil
-}
-
-func (c *Cluster) getAvailableMonNodes() ([]v1.Node, *v1.NodeList, error) {
-	nodeOptions := metav1.ListOptions{}
-	nodeOptions.TypeMeta.Kind = "Node"
-	nodes, err := c.context.Clientset.CoreV1().Nodes().List(nodeOptions)
-	if err != nil {
-		return nil, nil, err
-	}
-	logger.Debugf("there are %d nodes available for %d mons", len(nodes.Items), len(c.clusterInfo.Monitors))
-
-	// get the nodes that have mons assigned
-	nodesInUse, err := c.getNodesWithMons(nodes)
-	if err != nil {
-		logger.Warningf("could not get nodes with mons. %+v", err)
-		nodesInUse = util.NewSet()
-	}
-
-	// choose nodes for the new mons that don't have mons currently
-	availableNodes := []v1.Node{}
-	for _, node := range nodes.Items {
-		if !nodesInUse.Contains(node.Name) {
-			valid, err := k8sutil.ValidNode(node, c.placement)
-			if err != nil {
-				logger.Warning("failed to validate node %s %v", node.Name, err)
-			} else if valid {
-				availableNodes = append(availableNodes, node)
-			}
-		}
-	}
-
-	return availableNodes, nodes, nil
-}
-
-func (c *Cluster) getNodesWithMons(nodes *v1.NodeList) (*util.Set, error) {
-	// get the mon pods and their node affinity
-	options := metav1.ListOptions{LabelSelector: fmt.Sprintf("app=%s", appName)}
-	pods, err := c.context.Clientset.CoreV1().Pods(c.Namespace).List(options)
-	if err != nil {
-		return nil, err
-	}
-	nodesInUse := util.NewSet()
-	for _, pod := range pods.Items {
-		hostname := pod.Spec.NodeSelector[apis.LabelHostname]
-		logger.Debugf("mon pod on node %s", hostname)
-		name, ok := getNodeNameFromHostname(nodes, hostname)
-		if !ok {
-			logger.Errorf("mon %s on hostname %s not found in node list", pod.Name, hostname)
-		}
-		nodesInUse.Add(name)
-	}
-	return nodesInUse, nil
-}
-
-// Look up the immutable node name from the hostname label
-func getNodeNameFromHostname(nodes *v1.NodeList, hostname string) (string, bool) {
-	for _, node := range nodes.Items {
-		if node.Labels[apis.LabelHostname] == hostname {
-			return node.Name, true
-		}
-		if node.Name == hostname {
-			return node.Name, true
-		}
-	}
-	return "", false
 }
 
 var updateDeploymentAndWait = k8sutil.UpdateDeploymentAndWait
